better, nearly working (small number of broken MCUs)

+++ types.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct MCU {
-	int Y[64];
-	int Cb[64];
-	int Cr[64];
+	double Y[64];
+	double Cb[64];
+	double Cr[64];
 } typedef MCU;
 
 struct YCbCr {
